app/pyproject.toml
[tool.poetry]
name = "django-start-template"
version = "0.1.0"
description = ""
authors = ["hsuanchi <a0025071@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.5"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
djangorestframework = "^3.12.2"
django-q = "^1.3.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
